chore: Split up file structure

+++ src/color/color.c
@@ -6,7 +6,7 @@
 /*   By: ldulling <ldulling@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/02 11:45:02 by ldulling          #+#    #+#             */
-/*   Updated: 2023/12/14 13:01:53 by ldulling         ###   ########.fr       */
+/*   Updated: 2023/12/14 18:26:15 by ldulling         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,8 +37,8 @@ int	save_rgb_in_int(double smooth_color)
 	int	g;
 	int	b;
 
-	r = (int) (9 * (1 - smooth_color) * pow(smooth_color, 3) * 255);
-	g = (int) (15 * pow((1 - smooth_color), 2) * pow(smooth_color, 2) * 255);
-	b = (int) (8.5 * pow((1 - smooth_color), 3) * smooth_color * 255);
+	r = (9 * (1 - smooth_color) * pow(smooth_color, 3) * 255);
+	g = (15 * pow((1 - smooth_color), 2) * pow(smooth_color, 2) * 255);
+	b = (8.5 * pow((1 - smooth_color), 3) * smooth_color * 255);
 	return ((r << 16) | (g << 8) | b);
 }
